triangle type now works for double too and exception for empty reverse string

+++ assignment1/src/main/java/com/infosys/assignment1/service/ApiInterface.java
@@ -9,7 +9,7 @@ public interface ApiInterface{
 
 	String reverseWords(String s);
 
-	String triangleType(Integer a, Integer b, Integer c);
+	String triangleType(Double a, Double b, Double c);
 
 	LinkedHashMap<String, List<Integer>> makeOneArray(LinkedHashMap<String, List<Integer>> a);
